Refactored Symbolic Evaluator.
